+++ lib/yard/lint/validators/warnings/unknown_tag/config.rb
@@ -3,9 +3,7 @@
 module Yard
   module Lint
     module Validators
-      # Validators for checking YARD warnings
       module Warnings
-        # Validator for detecting unknown tags in documentation
         module UnknownTag
           # Configuration for UnknownTag validator
           class Config < ::Yard::Lint::Validators::Config